refactor: rename mcp package to mcpself
